model/requirements.txt

matplotlib
numpy
Pillow
torch
torchvision
tqdm
wandb
albumentations
segmentation_models_pytorch
pytorch-lightning==1.5.4